+++ src/admin/migrations/m190227_123549_cms_nav_item_strict_url_parsing.php
@@ -20,6 +20,6 @@ class m190227_123549_cms_nav_item_strict_url_parsing extends Migration
      */
     public function safeDown()
     {
-        $this->dropColumn('cms_nav_item', 'is_url_strict_parsing_disabled', $this->boolean());
+        $this->dropColumn('cms_nav_item', 'is_url_strict_parsing_disabled');
     }
 }